Fix MySQL parameter binding errors
- Fixed LIMIT parameter binding in MedicationSchedule.findByMedicationId by using template literal
- Changed frequency_hours column from INT to DECIMAL(10,2) to support decimal values (e.g., 0.5 hours)
- Added migration to alter existing frequency_hours column
- Updated initial table creation to use DECIMAL type

This fixes the 'Incorrect arguments to mysqld_stmt_execute' error when creating medications with decimal frequency hours.

+++ src/config/database.js
@@ -114,7 +114,7 @@ const runMigrations = async () => {
         name VARCHAR(255) NOT NULL,
         description TEXT,
         dosage VARCHAR(100),
-        frequency_hours INT NOT NULL,
+        frequency_hours DECIMAL(10,2) NOT NULL,
         start_time TIME NOT NULL,
         duration_days INT DEFAULT NULL,
         started_at DATE DEFAULT NULL,
